+++ XFrame/Core/Module/Container/ShareContainer.cs
@@ -68,7 +68,7 @@ namespace XFrame.Modules.Containers
         /// </summary>
         protected internal virtual void OnInit() { }
 
-        void IContainer.OnUpdate(float elapseTime)
+        void IContainer.OnUpdate(double elapseTime)
         {
             OnUpdate(elapseTime);
         }
@@ -77,7 +77,7 @@ namespace XFrame.Modules.Containers
         /// 更新生命周期
         /// </summary>
         /// <param name="elapseTime">逃逸时间</param>
-        protected internal virtual void OnUpdate(float elapseTime) { }
+        protected internal virtual void OnUpdate(double elapseTime) { }
 
         void IContainer.OnDestroy()
         {
