+++ XFrame/Core/Module/IUpdater.cs
@@ -10,6 +10,6 @@ namespace XFrame.Core
         /// 更新生命周期
         /// </summary>
         /// <param name="escapeTime">逃逸时间</param>
-        void OnUpdate(float escapeTime);
+        void OnUpdate(double escapeTime);
     }
 }
